@@ -5,19 +5,20 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "start":  "vue-cli-service serve"
+    "start": "vue-cli-service serve"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "bootstrap-vue": "^2.21.2",
+    "@vue/cli-service": "^5.0.8",
+    "axios": "^1.8.4",
+    "bootstrap": "^5.3.7",
+    "bootstrap-vue-next": "^0.30.4",
     "core-js": "^3.6.5",
-    "socket.io-client": "^4.2.0",
-    "vue": "^2.6.11",
+    "mitt": "^3.0.1",
+    "socket.io-client": "^4.8.1",
+    "vue": "^3.5.18",
     "vue-axios": "^3.3.7",
-    "vue-router": "^3.2.0",
-    "vue-socket.io-extended": "^4.2.0",
-    "vuex": "^3.4.0",
-    "@vue/cli-service": "^5.0.8"
+    "vue-router": "^4.5.1",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^5.0.8",
