@@ -4,9 +4,10 @@ from ..models import User, db
 from .util import LEADERBOARD_FIELDS
 from flask.json import jsonify
 
+
 class LeaderBoardResource(Resource):
-	@marshal_with(LEADERBOARD_FIELDS)
-	def get(self):
-		users = db.session.query(User.username, User.score).all()
-		print(users)
-		return users
+    @marshal_with(LEADERBOARD_FIELDS)
+    def get(self):
+        users = db.session.query(User.username, User.score).all()
+        print(users)
+        return users
