@@ -1,35 +1,26 @@
 from flask_restful import fields, reqparse
 
-USER_FIELDS = {
-	"id": fields.Integer,
-	"username": fields.String,
-}
-
 LEADERBOARD_FIELDS = {
-	"username": fields.String,
+    "username": fields.String,
     "score": fields.Integer,
 }
 
 ROOMS_FIELDS = {
-	"id": fields.Integer,
-	"name": fields.String,
-	"master": fields.Nested(USER_FIELDS),
-	"createdAt": fields.DateTime,
-	"modifiedAt": fields.DateTime,
-	"governmentType": fields.String
+    "id": fields.Integer,
+    "name": fields.String,
+    "startedAt": fields.DateTime(attribute="startedAt"),
+    "endedAt": fields.DateTime(attribute="endedAt"),
+    "governmentType": fields.String(
+        attribute=lambda x: x.government.type if x.government else None
+    ),
+    # Optionally, add a nested government field for members, etc.
 }
 
-JOIN_ROOM_FIELDS = {	
-	"roomId": fields.Integer
-}
+JOIN_ROOM_FIELDS = {"roomId": fields.Integer}
 
-LEAVE_ROOM_FIELDS = {
-	"roomId": fields.Integer
-}
+LEAVE_ROOM_FIELDS = {"roomId": fields.Integer}
 
-ROOM_WAITING_FIELDS = {
-	"user_id"
-}
+ROOM_WAITING_FIELDS = {"user_id"}
 
 roomParser = reqparse.RequestParser()
 roomParser.add_argument("name", type=str, required=True)
