@@ -3,7 +3,8 @@ from . import db, jwt
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime, timedelta
 
-rooms = db.Table("rooms", db.Column("roomId", db.Integer, db.ForeignKey("room.id"))),
+rooms = (db.Table("rooms", db.Column("roomId", db.Integer, db.ForeignKey("room.id"))),)
+
 
 class User(db.Model):
     __tablename__ = "user"
@@ -11,11 +12,13 @@ class User(db.Model):
     username = db.Column(db.String(60), index=True, unique=True)
     score = db.Column(db.Integer, default=0)
     passwordHash = db.Column(db.String(128))
-    rooms = db.relationship("Room", secondary="room_participants", lazy="subquery", viewonly=True)
+    rooms = db.relationship(
+        "Room", secondary="room_participants", lazy="subquery", viewonly=True
+    )
 
     @property
     def password(self):
-        raise AttributeError("You stupid fucking idiot")
+        raise AttributeError("You stupid idiot")
 
     @password.setter
     def password(self, newPassword):
@@ -24,14 +27,17 @@ class User(db.Model):
     def verifyPassword(self, testPassword):
         return check_password_hash(self.passwordHash, testPassword)
 
+
 @jwt.user_identity_loader
 def user_identity_lookup(user):
     return user.id
+
 
 @jwt.user_lookup_loader
 def user_lookup_callback(jwt_header, jwt_data):
     identity = jwt_data["sub"]
     return User.query.filter_by(id=identity).one_or_none()
+
 
 class Room(db.Model):
     __tablename__ = "room"
@@ -39,34 +45,82 @@ class Room(db.Model):
     name = db.Column(db.String(60), nullable=False)
     participants = db.relationship("RoomParticipants", lazy="subquery")
     startedAt = db.Column(db.DateTime, default=datetime.utcnow)
-    endedAt = db.Column(db.DateTime, default=datetime.utcnow)#TO-DO
-    #government = db.relationship("Government", lazy="subquery")
+    endedAt = db.Column(db.DateTime, default=datetime.utcnow)
+    government = db.relationship(
+        "Government", backref="room", uselist=False, lazy="subquery"
+    )
+
 
 class RoomParticipants(db.Model):
     __tablename__ = "room_participants"
     roomId = db.Column(db.Integer, db.ForeignKey("room.id"), primary_key=True)
     userId = db.Column(db.Integer, db.ForeignKey("user.id"), primary_key=True)
-    timeBank = db.Column(db.DateTime, default=datetime.min)#Have add or remove button
     clock = db.Column(db.DateTime, default=datetime.min + timedelta(days=1))
-    interest_rate = db.Column(db.Float, nullable=False)
-
-    #job = db.relationship("Job", lazy="subquery")
+    job_id = db.Column(db.Integer, db.ForeignKey("job.id"))
+    job = db.relationship("Job")
+    bleed = db.Column(db.Integer, default=0)
+    user = db.relationship("User")
 
 class Job(db.Model):
     __tablename__ = "job"
     id = db.Column(db.Integer, primary_key=True)
     tier = db.Column(db.Integer)
     name = db.Column(db.String(60), nullable=False)
-    description = db.Column(db.String(), nullable=False)    
+    type = db.Column(db.String(), nullable=False)
+
 
 class Government(db.Model):
     __tablename__ = "government"
     id = db.Column(db.Integer, primary_key=True)
     type = db.Column(db.String(60), default="Democracy")
-    description = db.Column(db.String(), nullable=False)
+    room_id = db.Column(db.Integer, db.ForeignKey("room.id"))
+    members = db.relationship(
+        "GovernmentMember", back_populates="government", lazy="subquery"
+    )
 
-class Laws(db.Model):
-    __tablename__ = "laws"
+
+class GovernmentMember(db.Model):
+    __tablename__ = "government_member"
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(), nullable=False)
-    description = db.Column(db.String(), nullable=False)
+    government_id = db.Column(db.Integer, db.ForeignKey("government.id"))
+    user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
+    role = db.Column(
+        db.String(32), nullable=False
+    )  # e.g. "leader", "advisor", "politburo"
+    government = db.relationship("Government", back_populates="members")
+    user = db.relationship("User")
+
+
+def serialize_room(room):
+    return {
+        "id": room.id,
+        "name": room.name,
+        # Use room.id as owner if that's your convention (no created_by/owner_id columns)
+        "owner_id": room.id,
+        "government": serialize_government(room.government),
+        "participants": [
+            {
+                "user_id": rp.userId,
+                "username": (rp.user.username if hasattr(rp, "user") and rp.user else None),
+                "job_tier": rp.job.tier if rp.job else None,
+                "job_name": rp.job.name if rp.job else None,
+                "bleed": rp.bleed,
+            }
+            for rp in room.participants
+        ],
+    }
+
+
+def serialize_government(government):
+    if not government:        return None
+    return {
+        "type": getattr(government, "type", None),
+        "members": [
+            {
+                "user_id": gm.user_id,
+                "username": gm.user.username if gm.user else None,
+                "role": gm.role,
+            }
+            for gm in (government.members or [])
+        ],
+    }
