@@ -1,71 +1,86 @@
 <template>
   <div id="app">
-    <head>
-      <title>Gerrys Game</title>
-    </head>
-    <b-navbar type="dark" variant="dark">
-      <b-navbar-brand to="/home" href="#">Gerrys Game</b-navbar-brand>
-        <b-collapse id="nav-collapse" is-nav>
-        <b-navbar-nav>
-          <b-nav-item to="/rooms">Active games</b-nav-item>
-          <template v-if="username != null && roomId != null"> 
-            <b-nav-item to="/room">Current Game</b-nav-item>
-          </template>
-        
-          <!--<b-nav-item to="/me/quizzes">My quizzes</b-nav-item> -->
-        </b-navbar-nav>
-        <b-navbar-nav class="ml-auto">
-          <template v-if="username != null">
-            <b-nav-item-dropdown right>
-              <template v-slot:button-content>
-                <em>{{ username }}</em>
-              </template>
-              <b-dropdown-item to="/me">Profile</b-dropdown-item>
-              <b-dropdown-item v-on:click="onLogout()">Sign Out</b-dropdown-item>
-            </b-nav-item-dropdown>
-          </template>
-          <template v-else>
-            <b-nav-item to="/login">Login/Register</b-nav-item>
-          </template>
-        </b-navbar-nav>
-      </b-collapse>
-    </b-navbar>
-    <b-container fluid>
-      <router-view/>
-    </b-container>
+    <BNavbar type="dark" variant="dark">
+      <BNavbarBrand>
+        <router-link to="/home" class="nav-link text-light">Gerrys Game</router-link>
+      </BNavbarBrand>
+      <BCollapse id="nav-collapse" is-nav>
+        <BNavbarNav>
+          <BNavItem>
+            <router-link to="/rooms" class="nav-link text-light">Active games</router-link>
+          </BNavItem>
+          <BNavItem v-if="roomId">
+            <router-link to="/room" class="nav-link text-light">Room</router-link>
+          </BNavItem>
+        </BNavbarNav>
+        <BNavbarNav class="ms-auto">
+          <BNavItem v-if="!username">
+            <router-link to="/login" class="nav-link text-light">Login</router-link>
+          </BNavItem>
+          <BNavItemDropdown
+            v-else
+            :text="username"
+            right
+            class="ml-2"
+            menu-class="bg-dark"
+            toggle-class="text-light"
+          >
+            <BDropdownItem @click="goToProfile">View Profile</BDropdownItem>
+            <BDropdownItem @click="onLogout">Logout</BDropdownItem>
+          </BNavItemDropdown>
+        </BNavbarNav>
+      </BCollapse>
+    </BNavbar>
+    <router-view />
   </div>
 </template>
+
 <script>
+import { BNavItemDropdown, BDropdownItem } from 'bootstrap-vue-next'
 
 export default {
   name: "App",
-  mounted() {
-     document.title = "Gerrys Game";
+  components: {
+    BNavItemDropdown,
+    BDropdownItem
   },
-  data() {
-		return {
-			name: 'RoomList',
-		}
-	},
-    methods: {
-      async onLogout () {
-        await this.$store.dispatch('logout').then(() => {
-          this.$router.push('/login')
-      })
-    }  
+  mounted() {
+    document.title = "Gerrys Game";
+  },
+  computed: {
+    username() {
+      return this.$store.state.auth.username;
     },
-    computed: {
-      username () {
-        return this.$store.state.auth.username
-      },
-      roomId () {
-        return this.$store.state.auth.roomId
-      },
-  }
-}
+    roomId() {
+      return this.$store.state.auth.roomId;
+    },
+  },
+  methods: {
+    goToProfile() {
+      this.$router.push({ name: 'Profile' }); // Change 'Profile' to your actual profile route name if needed
+    },
+    async onLogout() {
+      await this.$store.dispatch("auth/logout");
+      this.$router.push("/login");
+    },
+  },
+};
 </script>
+
 <style>
-  .navbar {
-    margin-bottom: 50px;
-  }
+.navbar {
+  margin-bottom: 50px;
+}
+
+/* Make dropdown menu text white */
+.bg-dark .dropdown-item,
+.bg-dark .dropdown-item:active,
+.bg-dark .dropdown-item:focus,
+.bg-dark .dropdown-item:hover {
+  color: #fff !important;
+  background-color: #343a40 !important;
+}
+.bg-dark .dropdown-menu {
+  background-color: #343a40 !important;
+}
 </style>
