@@ -1,5 +1,5 @@
 // Define your IP_ADDRESS
-export const IP_ADDRESS = "localhost";
+export const IP_ADDRESS = "3.11.22.254";
 
 // Use it to create API_URL
 export const API_URL = `http://${IP_ADDRESS}:5000/api`;
