@@ -1,57 +1,63 @@
 <template>
-    <b-form @submit="onSubmit">
-      <b-form-group
-        label="Username:"
-        description="Enter your username"
-        placeholder="username"
-        label-class="font-weight-bold pt-0">
-        <b-form-input
-          v-model="username"
-          required>
-          </b-form-input>
-      </b-form-group>
-      <b-form-group
-        label="Password:"
-        description="Enter your password"
-        placeholder="password"
-        label-class="font-weight-bold pt-0">
-        <b-form-input
-          v-model="password"
-          type="password"
-          required>
-          </b-form-input>
-      </b-form-group>
-      <div class="form-group">
-      <small v-if="errors" class="text-danger">{{errors}}</small>
-      </div>
-      <b-button type="submit" class="btn btn-success">Login</b-button>
-    </b-form>
+  <BForm @submit.prevent="onSubmit">
+    <BFormGroup label="Username" label-for="username-input">
+      <BFormInput
+        id="username-input"
+        v-model="username"
+        required
+        placeholder="Enter username"
+      />
+    </BFormGroup>
+    <BFormGroup label="Password" label-for="password-input">
+      <BFormInput
+        id="password-input"
+        v-model="password"
+        type="password"
+        required
+        placeholder="Enter password"
+      />
+    </BFormGroup>
+    <div class="mb-2">
+      <small v-if="error" class="text-danger">{{ error }}</small>
+    </div>
+    <BButton type="submit" variant="primary" :disabled="loading">
+      Login
+    </BButton>
+  </BForm>
 </template>
 
 <script>
+import { ref } from 'vue'
+import { useStore } from 'vuex'
+import { useRouter } from 'vue-router'
+import { BForm, BFormGroup, BFormInput, BButton } from 'bootstrap-vue-next'
+
 export default {
-  data () {
-    return {
-      username: "",
-      password: "",
+  components: { BForm, BFormGroup, BFormInput, BButton },
+  setup() {
+    const store = useStore()
+    const router = useRouter()
+    const username = ref('')
+    const password = ref('')
+    const error = ref('')
+    const loading = ref(false)
+
+    const onSubmit = () => {
+      error.value = ''
+      loading.value = true
+      store.dispatch('auth/login', { username: username.value, password: password.value })
+        .then(() => {
+          router.push({ name: 'Home' })
+        })
+        .catch(() => {
+          error.value = store.state.auth.errors || 'Login failed'
+        })
+        .finally(() => {
+          loading.value = false
+        })
     }
-  },
-  methods: {
-    onSubmit () {
-      var credentials = {
-        username: this.username,
-        password: this.password
-      }
-      this.$store.dispatch("login", credentials)
-      .then(() => {
-        this.$router.push("/home")
-      })
-    }
-  },
-  computed: {
-    errors () {
-      return this.$store.state.auth.errors
-    }
-}
+
+    return { username, password, error, loading, onSubmit }
+  }
 }
 </script>
